wip converting to use decimals instead of doubles

+++ Source/CalcEngine.Tests/BasicExcelFunctionTests.cs
@@ -9,14 +9,14 @@ namespace CalcEngine.Tests
         public void supported_functions()
         {
             CalcEngine calcEngine = new CalcEngine();
-            calcEngine.Test("=SUM(1,2,3)", 6);
-            calcEngine.Test("=SUM(6)", 6);
-            calcEngine.Test("=ABS(-1)", 1);
-            calcEngine.Test("=ABS(1)", 1);
-            calcEngine.Test("=AVERAGE(1,2)", 1.5);
-            calcEngine.Test("CEILING(1.8)", Math.Ceiling(1.8));
-            calcEngine.Test("=COUNT(1,2,3)", 3);
-            calcEngine.Test("=ROUND(1,1)", 1);
+            calcEngine.Test("=SUM(1,2,3)", 6m);
+            calcEngine.Test("=SUM(6)", 6m);
+            calcEngine.Test("=ABS(-1)", 1.0m);
+            calcEngine.Test("=ABS(1)", 1m);
+            calcEngine.Test("=AVERAGE(1,2)", 1.5m);
+            calcEngine.Test("CEILING(1.8)", (decimal)Math.Ceiling(1.8));
+            calcEngine.Test("=COUNT(1,2,3)", 3m);
+            calcEngine.Test("=ROUND(1,1)", 1m);
         }
 
         [Fact(Skip = "Should not be run, this if only done for doc purpose")]
