+++ Source/CalcEngine.Tests/ComplexExcelFunctionTests.cs
@@ -13,7 +13,7 @@ namespace CalcEngine.Tests
             calcEngine.Test("=SUM(1,2,AVERAGE(1,2) + 3)", 7.5);
         }
 
-        public void Eval(string expression, double expectedValue)
+        public void Eval(string expression, decimal expectedValue)
         {
             calcEngine.Test(expression, expectedValue);
         }
@@ -21,43 +21,43 @@ namespace CalcEngine.Tests
         [Fact]
         public void EvaluatesSumFunction()
         {
-            Eval("=SUM(1,2,3,4,5)", 15.0d);
-            Eval("=SUM(10,5)", 15.0d);
-            Eval("=SUM(1)", 1d);
-            Assert.Throws<Exception>(() => Eval("=SUM()", 0.0d));
+            Eval("=SUM(1,2,3,4,5)", 15.0m);
+            Eval("=SUM(10,5)", 15.0m);
+            Eval("=SUM(1)", 1m);
+            Assert.Throws<Exception>(() => Eval("=SUM()", 0.0m));
         }
 
         [Fact]
         public void EvaluatesAverageFunction()
         {
-            Eval("=AVERAGE(1,2,3,4,5)", 3.0d);
-            Eval("=AVERAGE(10,5)", 7.5d);
-            Eval("=AVERAGE(1)", 1d);
-            Assert.Throws<Exception>(() => Eval("=AVERAGE()", 0d));
+            Eval("=AVERAGE(1,2,3,4,5)", 3.0m);
+            Eval("=AVERAGE(10,5)", 7.5m);
+            Eval("=AVERAGE(1)", 1m);
+            Assert.Throws<Exception>(() => Eval("=AVERAGE()", 0m));
         }
 
         [Fact]
         public void EvaluatesRoundFunction()
         {
-            Eval("=ROUND(1.20,2)", 1.2d);
-            Eval("=ROUND(1.20)", 1d);
-            Eval("=ROUND(1.50)", 2d);
-            Assert.Throws<Exception>(() => Eval("=ROUND()", 0d));
+            Eval("=ROUND(1.20,2)", 1.2m);
+            Eval("=ROUND(1.20)", 1m);
+            Eval("=ROUND(1.50)", 2m);
+            Assert.Throws<Exception>(() => Eval("=ROUND()", 0m));
         }
 
         [Fact]
         public void EvaluatesWEIGHTEDFunction()
         {
-            Eval("=WEIGHTED(5, 5)", 25d);
-            Eval("=WEIGHTED(5, 0)", 0d);
-            Assert.Throws<Exception>(() => Eval("=WEIGHTED()", 0d));
-            Assert.Throws<Exception>(() => Eval("=WEIGHTED(1)", 0d));
+            Eval("=WEIGHTED(5, 5)", 25m);
+            Eval("=WEIGHTED(5, 0)", 0m);
+            Assert.Throws<Exception>(() => Eval("=WEIGHTED()", 0m));
+            Assert.Throws<Exception>(() => Eval("=WEIGHTED(1)", 0m));
         }
 
         [Fact]
         public void EvaluatesSumProductFunction()
         {
-            Eval("=ROUND(SUM(4*3,2*2,3*1)/SUM(4+2+3), 2)", 2.11d);
+            Eval("=ROUND(SUM(4*3,2*2,3*1)/SUM(4+2+3), 2)", 2.11m);
         }
 
         [Fact]
@@ -92,7 +92,7 @@ namespace CalcEngine.Tests
             Eval("=CEILING(2.45, 1)", 3);
             Eval("=CEILING(2.45, 2)", 4);
             Eval("=CEILING(-2.45, -2)", -4);
-            Eval("=CEILING(2.45, 0.2)", 2.6);
+            Eval("=CEILING(2.45, 0.2)", 2.6m);
 
             Assert.Throws<Exception>(() => Eval("=CEILING(2.45, -2)", 0));
         }
